Aggregation files restrict to all present

+++ Pipelines/geneanalysis/python/mod_onegenestitch.py
@@ -58,6 +58,8 @@ def run_pipeline(args):
     all_var_build = []
     
     # first concat all the dfs, variants and backgrount
+    exists_all_back = True
+    exists_all_var = True
     for pdb in pdbs:        
         # the file has already been turned into a dataframe called posscan_df.csv
         pdb_path = Paths.Paths(            
@@ -73,40 +75,54 @@ def run_pipeline(args):
         if exists(file_var_ps):
             fdf = FileDf.FileDf(file_var_ps)                        
             all_var_scan.append(fdf.openDataFrame())            
+        else:
+            exists_all_var = False
         if exists(file_var_bm):
             fdf = FileDf.FileDf(file_var_bm)                        
             all_var_build.append(fdf.openDataFrame())            
+        else:
+            exists_all_var = False
         if exists(file_back):
             fdf = FileDf.FileDf(file_back)            
             all_back.append(fdf.openDataFrame())
+        else:
+            exists_all_back = False
 
-    ddg_df_back = pd.concat(all_back,ignore_index=True)
-    ddg_df_var_scan = pd.concat(all_var_scan,ignore_index=True)
-    ddg_df_var_build = pd.concat(all_var_build,ignore_index=True)
-    ddg_df_back.to_csv(gene_path.gene_outputs+"ddg_background.csv", index=False)
-    ddg_df_var_scan.to_csv(gene_path.gene_outputs+"ddg_variant_ps.csv", index=False)
-    ddg_df_var_build.to_csv(gene_path.gene_outputs+"ddg_variant_bm.csv", index=False)
+    if exists_all_back:
+        ddg_df_back = pd.concat(all_back,ignore_index=True)
+        ddg_df_back.to_csv(gene_path.gene_outputs+"ddg_background.csv", index=False)
+        #DB Coverage reports    
+        plot_file = gene_path.gene_outputs + "ALL_coverage.png"
+        anav = Analysis.Analysis(ddg_df_back, gene)                        
+        anav.createPdbSummary(plot_file, "PDB Coverage")
+    else:
+        print("Gene stitch: not all pdb background files present for aggregation")
+    
+    if exists_all_var:
+        ddg_df_var_scan = pd.concat(all_var_scan,ignore_index=True)
+        ddg_df_var_build = pd.concat(all_var_build,ignore_index=True)    
+        ddg_df_var_scan.to_csv(gene_path.gene_outputs+"ddg_variant_ps.csv", index=False)
+        ddg_df_var_build.to_csv(gene_path.gene_outputs+"ddg_variant_bm.csv", index=False)
+    else:
+        print("Gene stitch: not all pdb variant files present for aggregation")
             
-    #DB Coverage reports    
-    plot_file = gene_path.gene_outputs + "ALL_coverage.png"
-    anav = Analysis.Analysis(ddg_df_back, gene)                        
-    anav.createPdbSummary(plot_file, "PDB Coverage")
-        
+            
     tags = ["SMHOM","SMEXP","AF","EXP"]
     for tag in tags:
         df_back = ddg_df_back.query("source=='"+tag+"'")
         df_scan = ddg_df_var_scan.query("source=='"+tag+"'")
         df_build = ddg_df_var_build.query("source=='"+tag+"'")        
         
-        anav_back = Analysis.Analysis(df_back, gene)
-        anav_back.createDdgResidue(gene_path.gene_outputs+tag+ "_background_rid.png", "Background (residue) "+ tag, xax="pdb_rid", dropnagene=False)
-        anav_back.createDdgResidue(gene_path.gene_outputs+tag+ "_background_gene.png", "Background (gene no) "+ tag, xax="gene_no", dropnagene=True)
+        if exists_all_back:
+            anav_back = Analysis.Analysis(df_back, gene)
+            anav_back.createDdgResidue(gene_path.gene_outputs+tag+ "_background_rid.png", "Background (residue) "+ tag, xax="pdb_rid", dropnagene=False)
+            anav_back.createDdgResidue(gene_path.gene_outputs+tag+ "_background_gene.png", "Background (gene no) "+ tag, xax="gene_no", dropnagene=True)
 
-        anav_scan = Analysis.Analysis(df_scan, gene)
-        anav_scan.createDdgResidue(gene_path.gene_outputs+tag+ "_variant_ps.png", "Variant PosScan "+ tag, xax="gene_no", dropnagene=False)
-
-        anav_build = Analysis.Analysis(df_build, gene)
-        anav_build.createDdgResidue(gene_path.gene_outputs+tag+ "_variant_bm.png", "Variant BuildModel "+ tag, xax="gene_no", dropnagene=False)
+        if exists_all_var:
+            anav_scan = Analysis.Analysis(df_scan, gene)
+            anav_scan.createDdgResidue(gene_path.gene_outputs+tag+ "_variant_ps.png", "Variant PosScan "+ tag, xax="gene_no", dropnagene=False)
+            anav_build = Analysis.Analysis(df_build, gene)
+            anav_build.createDdgResidue(gene_path.gene_outputs+tag+ "_variant_bm.png", "Variant BuildModel "+ tag, xax="gene_no", dropnagene=False)
                             
     print("### COMPLETED GENE STITCH job ###")
     print("MUTEIN SCRIPT ENDED")
